requirements.txt
pydantic~=2.11.9
requests
python-owasp-zap-v2.4
